Update runs to error if it runs twice

+++ libbids/run.py
@@ -56,6 +56,9 @@ class Run(Entity):
 
     def initialize_event_file(self) -> None:
         """Initializes the event file for writing"""
+        if self.event_file_path.exists():
+            raise Exception("Run data is already saved, please create a new run")
+
         self.eventbuf = open(self.event_filepath, "w")
         header: str = "\t".join(["onset", "duration", "trial_type"]) + "\n"
         self.eventbuf.writelines([header])
